+++ arduino/Shell.cpp
@@ -1,6 +1,7 @@
 #include "Shell.h"
 #include "Context.h"
 #include "Buzzer.h"
+#include "OutMux.h"
 #include "src/PunchCard.h"
 #include <RTClib.h>
 
@@ -8,15 +9,11 @@ namespace {
 
 const uint8_t MAX_SIZE = 32;
 
-void Prompt()
-{
-    Serial.print(F("Arduin-o-punch> "));
-}
-
 } // namespace;
 
-Shell::Shell(Context &context, Buzzer &buzzer)
-    : _context{context}
+Shell::Shell(OutMux &_outMux, Context &context, Buzzer &buzzer)
+    : _outMux{_outMux}
+    , _context{context}
     , _buzzer{buzzer}
 {
     _buffer.reserve(MAX_SIZE);
@@ -24,32 +21,54 @@ Shell::Shell(Context &context, Buzzer &buzzer)
 
 void Shell::Setup()
 {
-    Serial.println();
-    Prompt();
+    _outMux.println();
+    _PrintPrompt();
+}
+
+void Shell::_PrintPrompt()
+{
+    _outMux.print(F("Arduin-o-punch> "));
+}
+
+void Shell::_ProcessChar(char ch)
+{
+    if (!_buffer.length()) {
+        // In the automated communication we don't need the echo, neither the prompt.
+        // If timeout elapses since the first character, revert to interactive move.
+        _echo_timeout = millis() + 100;
+        _echo_idx = 0;
+    }
+    _buffer += ch;
+    if (_buffer.length() >= MAX_SIZE || _buffer.endsWith("\r")) {
+        // Echo the rest of the command if necessary
+        if (Tick())
+            _outMux.println();
+        _Process();
+        _buffer.remove(0, _buffer.length());
+        if (Tick())
+            _PrintPrompt();
+    }
 }
 
 void Shell::OnSerial()
 {
     while (Serial.available()) {
-        char ch = Serial.read();
-        if (!_buffer.length()) {
-            // In the automated communication we don't need the echo, neither the prompt.
-            // If timeout elapses since the first character, revert to interactive move.
-            _echo_timeout = millis() + 100;
-            _echo_idx = 0;
-        }
-        _buffer += ch;
-        if (_buffer.length() >= MAX_SIZE || _buffer.endsWith("\r")) {
-            // Echo the rest of the command if necessary
-            if (Tick())
-                Serial.println();
-            _Process();
-            _buffer.remove(0, _buffer.length());
-            if (Tick())
-                Prompt();
-        }
+        int ch = Serial.read();
+        if (ch >= 0)
+            _ProcessChar(ch);
     }
 }
+
+#ifdef ESP32
+
+void Shell::ProcessInput(const uint8_t *data, int size)
+{
+    while (size--) {
+        _ProcessChar(*data++);
+    }
+}
+
+#endif //ESP32
 
 // Return true if started echoing (interactive mode)
 boolean Shell::Tick()
@@ -57,53 +76,53 @@ boolean Shell::Tick()
     if (_echo_timeout > millis())
         return false;
     while (_echo_idx < _buffer.length())
-        Serial.print(_buffer[_echo_idx++]);
+        _outMux.print(_buffer[_echo_idx++]);
     return true;
 }
 
 void Shell::_Process()
 {
     if (_buffer.startsWith(F("help"))) {
-        Serial.println(F("Commands:"));
-        Serial.println(F("info              All info"));
-        Serial.println(F("id                ID"));
-        Serial.print  (F("id 33             Set ID: check="));
-        Serial.print(AOP::PunchCard::CHECK_STATION);
-        Serial.print(F(" start="));
-        Serial.print(AOP::PunchCard::START_STATION);
-        Serial.print(F(" finish="));
-        Serial.println(AOP::PunchCard::FINISH_STATION);
-        Serial.println(F("key               Key"));
-        Serial.println(F("key 112233445566  Set key"));
-        Serial.println(F("clock             Clock reading (ms)"));
-        Serial.println(F("clock 12345000    Set clock (ms)"));
-        Serial.println(F("date              Current date"));
-        Serial.println(F("time              Current time"));
-        Serial.println(F("timestamp         Print UNIX timestamp"));
-        Serial.println(F("timestamp 12345   Set date and time with UNIX timestamp"));
-        Serial.println(F("recfmt 256 2      Clear/prepare recorder (card count, bits per record)"));
-        Serial.println(F("rec               List punched cards"));
-        Serial.println(F("rec 123           Print punch count for a card"));
-        Serial.println(F("recclr 123        Clear card from the record"));
-        Serial.println(F("recdays           How many days to keep the record"));
-        Serial.println(F("recdays 1         Clear record after so many days"));
+        _outMux.println(F("Commands:"));
+        _outMux.println(F("info              All info"));
+        _outMux.println(F("id                ID"));
+        _outMux.print  (F("id 33             Set ID: check="));
+        _outMux.print(AOP::PunchCard::CHECK_STATION);
+        _outMux.print(F(" start="));
+        _outMux.print(AOP::PunchCard::START_STATION);
+        _outMux.print(F(" finish="));
+        _outMux.println(AOP::PunchCard::FINISH_STATION);
+        _outMux.println(F("key               Key"));
+        _outMux.println(F("key 112233445566  Set key"));
+        _outMux.println(F("clock             Clock reading (ms)"));
+        _outMux.println(F("clock 12345000    Set clock (ms)"));
+        _outMux.println(F("date              Current date"));
+        _outMux.println(F("time              Current time"));
+        _outMux.println(F("timestamp         Print UNIX timestamp"));
+        _outMux.println(F("timestamp 12345   Set date and time with UNIX timestamp"));
+        _outMux.println(F("recfmt 256 2      Clear/prepare recorder (card count, bits per record)"));
+        _outMux.println(F("rec               List punched cards"));
+        _outMux.println(F("rec 123           Print punch count for a card"));
+        _outMux.println(F("recclr 123        Clear card from the record"));
+        _outMux.println(F("recdays           How many days to keep the record"));
+        _outMux.println(F("recdays 1         Clear record after so many days"));
     } else if (_buffer.startsWith(F("info"))) {
-        Serial.println(F("version=2.1.0"));
-        Serial.print(F("id="));
+        _outMux.println(F("version=2.1.0"));
+        _outMux.print(F("id="));
         _PrintId();
-        Serial.print(F("key="));
+        _outMux.print(F("key="));
         _PrintKey();
         auto now = _context.GetDateTime();
-        Serial.print(F("date="));
+        _outMux.print(F("date="));
         _PrintDate(now);
-        Serial.print(F("time="));
+        _outMux.print(F("time="));
         _PrintTime(now);
-        Serial.print(F("rec="));
-        Serial.print(_context.GetRecorder().GetSize());
-        Serial.print(F(" x "));
-        Serial.print(static_cast<int>(_context.GetRecorder().GetBitsPerRecord()));
-        Serial.println(F(" bpr"));
-        Serial.print(F("recdays="));
+        _outMux.print(F("rec="));
+        _outMux.print(_context.GetRecorder().GetSize());
+        _outMux.print(F(" x "));
+        _outMux.print(static_cast<int>(_context.GetRecorder().GetBitsPerRecord()));
+        _outMux.println(F(" bpr"));
+        _outMux.print(F("recdays="));
         _PrintRecordRetainDays();
     } else if (_buffer.startsWith(F("id "))) {
         _SetId(_buffer.c_str() + 3);
@@ -142,8 +161,8 @@ void Shell::_Process()
     } else if (_buffer.startsWith(F("rec"))) {
         _RecorderList();
     } else {
-        Serial.print(F("Unknown command: "));
-        Serial.println(_buffer);
+        _outMux.print(F("Unknown command: "));
+        _outMux.println(_buffer);
     }
 }
 
@@ -199,9 +218,9 @@ void Shell::_PrintKey()
     for (uint8_t i = 0; i < Context::KEY_SIZE; ++i) {
         char buf[3];
         sprintf(buf, "%02X", key[i]);
-        Serial.print(buf);
+        _outMux.print(buf);
     }
-    Serial.println();
+    _outMux.println();
 }
 
 void Shell::_SetClock(const char *str)
@@ -211,7 +230,7 @@ void Shell::_SetClock(const char *str)
 
 void Shell::_PrintClock(const DateTime &time)
 {
-    Serial.println(_context.GetClock(&time));
+    _outMux.println(_context.GetClock(&time));
 }
 
 void Shell::_SetTimestamp(const char *str)
@@ -222,43 +241,39 @@ void Shell::_SetTimestamp(const char *str)
 
 void Shell::_PrintTimestamp()
 {
-    Serial.println(_context.GetDateTime().unixtime());
+    _outMux.println(_context.GetDateTime().unixtime());
 }
 
-namespace {
-
-void PrintDD(uint8_t d)
+void Shell::_PrintDD(uint8_t d)
 {
     if (d <= 9)
-        Serial.print(F("0"));
-    Serial.print(d);
+        _outMux.print(F("0"));
+    _outMux.print(d);
 }
-
-} //namespace;
 
 void Shell::_PrintDate(const DateTime &date)
 {
-    Serial.print(date.year());
-    Serial.print(F("-"));
-    PrintDD(date.month());
-    Serial.print(F("-"));
-    PrintDD(date.day());
-    Serial.println();
+    _outMux.print(date.year());
+    _outMux.print(F("-"));
+    _PrintDD(date.month());
+    _outMux.print(F("-"));
+    _PrintDD(date.day());
+    _outMux.println();
 }
 
 void Shell::_PrintTime(const DateTime &time)
 {
-    PrintDD(time.hour());
-    Serial.print(F(":"));
-    PrintDD(time.minute());
-    Serial.print(F(":"));
-    PrintDD(time.second());
-    Serial.println();
+    _PrintDD(time.hour());
+    _outMux.print(F(":"));
+    _PrintDD(time.minute());
+    _outMux.print(F(":"));
+    _PrintDD(time.second());
+    _outMux.println();
 }
 
 void Shell::_PrintId()
 {
-    Serial.println(_context.GetId());
+    _outMux.println(_context.GetId());
 }
 
 void Shell::_SetId(const char *str)
@@ -273,14 +288,14 @@ void Shell::_RecorderFormat(const char *str)
 
     auto res = _context.GetRecorder().Format(count, bits_per_record, _context.GetDateTime().unixtime());
     if (res < 0) {
-        Serial.print(F("Error "));
-        Serial.println(-res);
+        _outMux.print(F("Error "));
+        _outMux.println(-res);
     } else {
-        Serial.println(F("OK"));
-        Serial.print(F("count="));
-        Serial.print(_context.GetRecorder().GetSize());
-        Serial.print(F(" bits_per_record="));
-        Serial.println(_context.GetRecorder().GetBitsPerRecord());
+        _outMux.println(F("OK"));
+        _outMux.print(F("count="));
+        _outMux.print(_context.GetRecorder().GetSize());
+        _outMux.print(F(" bits_per_record="));
+        _outMux.println(_context.GetRecorder().GetBitsPerRecord());
     }
 }
 
@@ -288,37 +303,40 @@ void Shell::_RecorderCheck(const char *str)
 {
     uint16_t card_id = ParseNum<uint16_t>(str);
     uint16_t punch_count = _context.GetRecorder().GetRecordCount(card_id);
-    Serial.println(punch_count);
+    _outMux.println(punch_count);
 }
 
 void Shell::_RecorderClear(const char *str)
 {
     uint16_t card_id = ParseNum<uint16_t>(str);
     auto ok = _context.GetRecorder().Record(card_id, -1);
-    Serial.println(ok ? F("FAIL") : F("OK"));
+    _outMux.println(ok ? F("FAIL") : F("OK"));
 }
 
 void Shell::_RecorderList()
 {
-    Serial.print(F("Size="));
-    Serial.println(_context.GetRecorder().GetSize());
+    _outMux.print(F("Size="));
+    _outMux.println(_context.GetRecorder().GetSize());
     struct Printer : AOP::Recorder::IVisitor
     {
+        OutMux &_outMux;
+        Printer(OutMux &outMux) : _outMux{outMux} { }
+
         void OnCard(uint16_t card, uint8_t count, void *ctx) override
         {
-            Serial.print(' ');
-            Serial.print(card);
-            Serial.print(':');
-            Serial.print(static_cast<uint16_t>(count));
+            _outMux.print(' ');
+            _outMux.print(card);
+            _outMux.print(':');
+            _outMux.print(static_cast<uint16_t>(count));
         }
-    } printer;
+    } printer{_outMux};
     _context.GetRecorder().List(printer, &printer);
-    Serial.println();
+    _outMux.println();
 }
 
 void Shell::_PrintRecordRetainDays()
 {
-    Serial.println(_context.GetRecordRetainDays());
+    _outMux.println(_context.GetRecordRetainDays());
 }
 
 void Shell::_SetRecordRetainDays(const char *str)